requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.7.*
pydantic-settings==2.3.*
sqlalchemy==2.0.*
alembic==1.13.*
psycopg[binary]==3.2.*
python-jose[cryptography]==3.3.0.*